rekee

Module edition

Set of tiles that belong to the same game edition (core box or expansion).

(Sidenote: I have not fully settled on the term “edition” yet; “box” might be better but is already widely used for a different purpose in the Rust std library. This type name might still change until the Rekee library reaches version 1.0)

Enums

Edition

Rallyman game edition (core box or one of the expansions).

ParseEditionError

ParseSeriesError

Series

Rallyman game series (GT or DIRT).