rekee::edition

Enum Edition

pub enum Edition {
    GtCoreBox,
    GtChampionship,
    GtWorldTour,
    GtTeamChallenge,
    GtAdrenalinePack,
    DirtCoreBox,
    DirtMonteCarlo,
    Dirt110Percent,
    DirtRx,
    DirtClimb,
    DirtCopilotPack,
}

Rallyman game edition (core box or one of the expansions).

Variants

GtCoreBox

Rallyman: GT core box

GtChampionship

Championship expansion for Rallyman: GT

GtWorldTour

World Tour expansion for Rallyman: GT

GtTeamChallenge

Team Challenge expansion for Rallyman: GT

GtAdrenalinePack

Adrenaline Pack expansion for Rallyman: GT

DirtCoreBox

Rallyman: DIRT core box

DirtMonteCarlo

Monte Carlo expansion for Rallyman: DIRT

Dirt110Percent

110% expansion for Rallyman: DIRT

DirtRx

RX expansion for Rallyman: DIRT

DirtClimb

Climb expansion for Rallyman: DIRT

DirtCopilotPack

Copilot Pack expansion for Rallyman: DIRT

Implementations

impl Edition

pub fn all_tiles() -> Vec<TileId>

Returns all the tiles of the game joined into a sorted list.

Examples

let tiles = Edition::all_tiles();
assert_eq!(tiles.first(), Some(&tile!(101)));
assert_eq!(tiles.last(), Some(&tile!(905, b)));
assert_eq!(tiles.len(), 307);

pub fn gt_tiles() -> Vec<TileId>

๐Deprecated: Please use Series::editions() together with Edition::tiles() instead

Returns tiles part of Rallyman: GT game editions joined into a sorted list.

Examples

let tiles = Edition::gt_tiles();
assert_eq!(tiles.first(), Some(&tile!(101)));
assert_eq!(tiles.last(), Some(&tile!(901, b)));
assert_eq!(tiles.len(), 145);

pub fn dirt_tiles() -> Vec<TileId>

๐Deprecated: Please use Series::editions() together with Edition::tiles() instead

Returns tiles part of Rallyman: DIRT game editions joined into a sorted list.

Examples

let tiles = Edition::dirt_tiles();
assert_eq!(tiles.first(), Some(&tile!(201, a)));
assert_eq!(tiles.last(), Some(&tile!(905, b)));
assert_eq!(tiles.len(), 162);

pub fn tiles(&self) -> Vec<TileId>

Returns the tiles of a specific game edition.

Examples

let tiles = Edition::GtCoreBox.tiles();
assert_eq!(tiles.first(), Some(&tile!(101)));
assert_eq!(tiles.last(), Some(&tile!(120, b)));
assert_eq!(tiles.len(), 63);

pub fn contains_tile(&self, tile: TileId) -> bool

Check whether the edition contains a specific tile.

Returns true when the identifier is found in the list of tiles.

Compares the full tile identifier, including the graphical variant. To compare the base identifier only use Edition::contains_base_tile() instead. See TileId::base() for more information on the difference between base and full tile identifiers.

Examples

let edition = Edition::DirtCoreBox;
assert_eq!(edition.contains_tile(tile!(201, a)), true);
assert_eq!(edition.contains_tile(tile!(325, b)), false);

let edition = Edition::DirtClimb;
assert_eq!(edition.contains_tile(tile!(201, a)), false);
assert_eq!(edition.contains_tile(tile!(325, b)), true);

let edition = Edition::GtWorldTour;
assert_eq!(edition.contains_tile(tile!(124, a, 1)), false);
assert_eq!(edition.contains_tile(tile!(124, a, 2)), true);

pub fn contains_base_tile(&self, tile: TileId) -> bool

Check whether the edition matches a given base tile identifier.

Returns true when a tile with matching base identifier is found in the list of tiles.

Compares the base tile identifier, with graphical variant excluded. To compare the full identifier use Edition::contains_tile() instead. See TileId::base() for more information on the difference between base and full tile identifiers.

Examples

let edition = Edition::DirtCoreBox;
assert_eq!(edition.contains_base_tile(tile!(201, a)), true);
assert_eq!(edition.contains_base_tile(tile!(325, b)), false);

let edition = Edition::DirtClimb;
assert_eq!(edition.contains_base_tile(tile!(201, a)), false);
assert_eq!(edition.contains_base_tile(tile!(325, b)), true);

let edition = Edition::GtWorldTour;
assert_eq!(edition.contains_base_tile(tile!(124, a, 1).base()), true);
assert_eq!(edition.contains_base_tile(tile!(124, a, 2).base()), true);

pub fn contains(&self, tile: TileId) -> bool

๐Deprecated: Please use Edition::contains_tile() instead

pub fn contains_base(&self, tile: TileId) -> bool

๐Deprecated: Please use Edition::contains_base_tile() instead

pub fn contains_token(&self, token: TokenId) -> bool

Check whether the edition contains a specific token.

Returns true when the identifier belongs to the given edition.

Examples

let edition = Edition::DirtCopilotPack;
assert_eq!(edition.contains_token(TokenId::Chicane(Terrain::Gravel)), true);
assert_eq!(edition.contains_token(TokenId::Oxygen(1)), false);

let edition = Edition::DirtClimb;
assert_eq!(edition.contains_token(TokenId::Chicane(Terrain::Gravel)), false);
assert_eq!(edition.contains_token(TokenId::Oxygen(1)), true);

pub fn is_expansion(&self) -> bool

Check whether the edition is a stand-alone game, or an expansion module.

Examples

assert_eq!(Edition::GtAdrenalinePack.is_expansion(), true);
assert_eq!(Edition::GtCoreBox.is_expansion(), false);

pub fn is_discontinued(&self) -> bool

Check whether manufacturing of the edition has been stopped.

The 110% and Copilot Pack extensions for Rallyman: DIRT are considered discontinued, as the publisher has replaced them with the Monte Carlo expansion. The RX and Climb extensions are not considered discontinued, hopefully they will get a future reprint.

Examples

assert_eq!(Edition::DirtMonteCarlo.is_discontinued(), false);
assert_eq!(Edition::DirtCopilotPack.is_discontinued(), true);

pub fn series(&self) -> Series

Game series that the edition belongs to.

Examples

let series = Edition::GtWorldTour.series();
assert_eq!(series, Series::Gt);

let series = Edition::Dirt110Percent.series();
assert_eq!(series, Series::Dirt);

pub fn iter() -> Iter<'static, Self>

Iterator over all game editions.