rekee::hexagon

Enum Direction

pub enum Direction {
    A = 0,
    B = 1,
    C = 2,
    D = 3,
    E = 4,
    F = 5,
}

Direction within a hexagon grid.

Used to indicate one of the six possible moving directions within the grid and to specify the rotation angle of hexagon tiles.

Examples

let dir = Direction::A;
assert_eq!(dir, 0.into());
assert_eq!(dir + Direction::B, Direction::B);
assert_eq!(dir - Direction::B, Direction::F);

Variants

A = 0

Direction of the positive q axis.

B = 1

Direction of the positive r axis.

C = 2

Direction of the positive s axis.

D = 3

Direction of the negative q axis.

E = 4

Direction of the negative r axis.

F = 5

Direction of the negative s axis.

Implementations

impl Direction

pub fn iter() -> Iter<'static, Self>

Iterator over all six directions.

Examples

let mut dirs = Direction::iter();
assert_eq!(dirs.next(), Some(&Direction::A));
assert_eq!(dirs.next(), Some(&Direction::B));
let mut dirs = dirs.skip(3);
assert_eq!(dirs.next(), Some(&Direction::F));
assert_eq!(dirs.next(), None);

pub fn opposite(&self) -> Self

Get opposite direction (point reflection).

Examples

let dir = Direction::A;
assert_eq!(dir.opposite(), Direction::D);

let dir = Direction::F;
assert_eq!(dir.opposite(), Direction::C);

pub fn rotated_left(&self) -> Self

Get direction rotated left (counter-clockwise).

Examples

let dir = Direction::B;
assert_eq!(dir.rotated_left(), Direction::A);

let dir = Direction::A;
assert_eq!(dir.rotated_left(), Direction::F);

pub fn rotated_right(&self) -> Self

Get direction rotated right (clockwise).

Examples

let dir = Direction::A;
assert_eq!(dir.rotated_right(), Direction::B);

let dir = Direction::F;
assert_eq!(dir.rotated_right(), Direction::A);

pub fn to_angle(&self) -> f32

Convert this hexagon grid direction into an angle in degrees.

Examples

let dir = Direction::A;
assert_eq!(dir.to_angle(), 0.0);

let dir = Direction::B;
assert_eq!(dir.to_angle(), 60.0);