Struct rekee::hexagon::Coordinate

pub struct Coordinate { /* fields omitted */ }

Coordinate within a hexagon grid.

The two-dimensional hexagon grid consists of three axes q, r, and s, which are rotated by 60° to each other. A coordinate is stored sparse – s is omitted as it can be computed from q and r on demand.

Examples

let mut pos = Coordinate::new(1, 2);
pos = pos + (2, 3).into();
assert_eq!(pos.q(), 3);
assert_eq!(pos.r(), 5);
assert_eq!(pos.s(), -8);

Implementations

impl Coordinate

pub const fn new(q: i32, r: i32) -> Self

Creates a new grid coordinate from positions on q and r axis.

pub fn q(&self) -> i32

Coordinate position on q axis.

pub fn r(&self) -> i32

Coordinate position on r axis.

pub fn s(&self) -> i32

Coordinate position on s axis.

pub fn neighbor<T>(&self, dir: T) -> Self where
T: Into<Direction>,

Get adjacent grid coordinate based on the given direction.

Examples

let pos = Coordinate::new(1, 0);
assert_eq!(pos.neighbor(Direction::A), (2, 0).into());
assert_eq!(pos.neighbor(Direction::B), (1, 1).into());

pub fn rotated_left(&self) -> Self

Get coordinate rotated left (counter-clockwise) by 60° around the grid center.

Examples

let pos = Coordinate::new(1, 0);
assert_eq!(pos.rotated_left(), (1, -1).into());

pub fn rotated_right(&self) -> Self

Get coordinate rotated right (clockwise) by 60° around the grid center.

Examples

let pos = Coordinate::new(1, 0);
assert_eq!(pos.rotated_right(), (0, 1).into());

pub fn to_pixel(&self, layout: &Layout) -> Point

Convert this hexagon grid coordinate into x/y pixel positions.

pub fn from_pixel_rounded(layout: &Layout, p: Point) -> Self

Convert x/y pixel positions back into a hexagon grid coordinate.