Struct rekee::hexagon::FloatCoordinate

pub struct FloatCoordinate { /* private fields */ }

Floating-point variant of a grid Coordinate.

Mainly used as an intermediate type when converting between x/y pixel positions and hexagon grid coordinates.

Implementations

impl FloatCoordinate

pub const fn new(q: f32, r: f32) -> Self

Creates a new floating-point grid coordinate from positions on q and r axis.

pub fn q(&self) -> f32

pub fn r(&self) -> f32

pub fn s(&self) -> f32

pub fn rotate<D>(self, dir: D) -> Selfwhere D: Into<FloatDirection>,

Rotate this hexagon grid coordinate by the given floating-point direction.

Examples

let pos = FloatCoordinate::new(1.0, 0.0)
    .rotate(FloatDirection(-1.0));

let check = FloatCoordinate::new(1.0, -1.0);
assert!((pos.q() - check.q()).abs() < 1e-6);
assert!((pos.r() - check.r()).abs() < 1e-6);

pub fn to_pixel(self, layout: &Layout) -> Point

Convert this floating-point grid coordinate into x/y pixel positions.

pub fn from_pixel(layout: &Layout, p: Point) -> Self

Convert x/y pixel positions into a floating-point grid coordinate.

pub fn to_coordinate(self) -> Coordinate

Convert this floating-point grid coordinate into the nearest hexagon grid coordinate.