Struct rekee::hexagon::Rect

pub struct Rect {
    pub left: f32,
    pub top: f32,
    pub width: f32,
    pub height: f32,
}

Rectangle area within a grid of rectangular pixels. It consists of the four values: top, left, width, and height. Both axes of the grid are aligned like the coordinate system of a web page: x runs from left to right, y runs from top to bottom.

Examples

let rect = Rect::new(1.0, 2.0, 4.0, 3.0);
assert_eq!(rect.left, 1.0);
assert_eq!(rect.top, 2.0);
assert_eq!(rect.right(), 5.0);
assert_eq!(rect.bottom(), 5.0);

Fields
left: f32top: f32width: f32height: f32
Implementations

impl Rect

pub fn new(left: f32, top: f32, width: f32, height: f32) -> Self

Creates a new rectangle from the left, top, width and height values. Values for width or height are allowed to be negative, the rectangle will be normalized in that case.

Examples

let rect = Rect::new(1.0, 2.0, 3.0, -4.0);
assert_eq!(rect.left, 1.0);
assert_eq!(rect.top, -2.0);
assert_eq!(rect.right(), 4.0);
assert_eq!(rect.bottom(), 2.0);

pub fn right(&self) -> f32

Right position of the rectangle.

pub fn bottom(&self) -> f32

Bottom position of the rectangle.

pub fn center(&self) -> Point

Center point of the rectangle.

Examples

let rect = Rect::new(0.0, 1.0, 3.0, 4.0);
assert_eq!(rect.center(), Point(1.5, 3.0));

pub fn top_left(&self) -> Point

Top-left corner of the rectangle.

Examples

let rect = Rect::new(0.0, 1.0, 3.0, 4.0);
assert_eq!(rect.top_left(), Point(0.0, 1.0));
assert_eq!(rect.left, 0.0);
assert_eq!(rect.top, 1.0);

pub fn bottom_right(&self) -> Point

Bottom-right corner of the rectangle.

Examples

let rect = Rect::new(0.0, 1.0, 3.0, 4.0);
assert_eq!(rect.bottom_right(), Point(3.0, 5.0));
assert_eq!(rect.right(), 3.0);
assert_eq!(rect.bottom(), 5.0);

pub fn union(&self, other: &Rect) -> Rect

Calculates the union of two rectangles, the smallest rectangle that fully contains both source rectangles.

Examples

let a = Rect::new(1.0, 2.0, 3.0, 4.0);
let b = Rect::new(-3.0, -4.0, 2.0, 1.0);
let rect = a.union(&b);
assert_eq!(rect.left, -3.0);
assert_eq!(rect.top, -4.0);
assert_eq!(rect.width, 7.0);
assert_eq!(rect.height, 10.0);

pub fn with_padding(&self, value: f32) -> Self

Creates a new rectangle from the existing one by adding a padding border on all four sides.

Examples

let a = Rect::new(1.0, 2.0, 3.0, 4.0);
let rect = a.with_padding(0.5);
assert_eq!(rect.left, 0.5);
assert_eq!(rect.top, 1.5);
assert_eq!(rect.width, 4.0);
assert_eq!(rect.height, 5.0);