Macro rekee::tile

macro_rules! tile {
    ($num:expr) => { ... };
    ($num:expr, a) => { ... };
    ($num:expr, b) => { ... };
    ($num:expr, a, $var:expr) => { ... };
    ($num:expr, b, $var:expr) => { ... };
}

Creates a new tile identifier.

The tile! macro adds convenience on directly calling TileId::new() as it allows to skip side or var parameters, if unused. Also the value for parameter side is restricted to a and b to avoid invalid tile identifiers.

Examples

let tile: TileId = tile!(102, a);
assert_eq!(tile.num(), 102);
assert_eq!(tile.side(), 1);
assert_eq!(tile.var(), 0);