rekee

Module tile

Tile identifier and characteristics.

Structs

DangerLevelSummary

Tile count information for a specific terrain danger level.

EditionSummary

Edition and tile count information for a specific edition.

GroupByDangerLevel

An iterator that yields the danger levels that are included in a list of tiles.

GroupByEdition

An iterator that yields the editions that are necessary to assemble a list of tiles.

GroupByTerrain

An iterator that yields the terrain surfaces that are included in a list of tiles.

TerrainSummary

Tile count information for a specific terrain surface.

TileId

Identifier of a game tile.

TileInfo

Information about tile characteristics like graphical variants, connections, and edge types.

TileSummary

Count information for a specific base tile.

Enums

Connection

Internal tile connection information. Describes how a specific tile edge is connected to other edges of the same tile.

ConnectionHint

Tile connection hint. Used to describe the orientation of adjacent tiles along the track in Map::append().

DangerLevel

Danger level information for a tile.

Edge

External tile edge information. Describes how the track spaces of the edge are connected to adjacent tiles.

Pacenote

ParseDangerLevelError

ParseHintError

ParseTerrainError

Terrain

Terrain surface information for a tile.

Traits

TileList

Utility methods on any list of TileIds.