Enum rekee::tile::Terrain

pub enum Terrain {
    None,
    Asphalt(u8),
    Gravel(u8),
    Snow(u8),
}

Tile terrain information. Describes the track surface of a tile together with the danger level.

The meaning of the danger level value is:

0: tile has no danger level assigned

1: tile has danger level “yellow” (low)

2: tile has danger level “orange” (medium)

3: tile has danger level “red” (high)

Use TileInfo::terrain() to get the terrain data of a specific tile.

Variants

None

Asphalt(u8)

Gravel(u8)

Snow(u8)

Implementations

impl Terrain

pub fn danger_level(&self) -> u8

Danger level of a tile.

Examples

let terrain = Terrain::Asphalt(3);
assert_eq!(terrain.danger_level(), 3);

let terrain = Terrain::None;
assert_eq!(terrain.danger_level(), 0);

pub fn surface(&self) -> Terrain

Surface of a tile.

Returns a copy of the terrain with danger level set to zero.

pub fn eq_surface(&self, other: Terrain) -> bool

Compares the surface of two tiles and ignores the danger level.

Examples

let terrain = Terrain::Asphalt(3);
assert!(terrain.eq_surface(Terrain::Asphalt(1)));

let terrain = Terrain::None;
assert!(!terrain.eq_surface(Terrain::Gravel(0)));