Struct rekee::tile::TileInfo

pub struct TileInfo { /* fields omitted */ }

Information about tile characteristics like graphical variants, connections, and edge types.

This tile characteristics data is used by Map to determine the best tile direction and next active position when adding a tile.

Use TileInfo::get() to lookup tile information by identifier. Note that the list of tiles is static, it is not possible to add or change any tile information entry at runtime.

Examples

let info = TileInfo::get(tile!(103, a, 2));
assert!(info.is_some());

Implementations

impl TileInfo

pub fn base_id(&self) -> TileId

Base identifier of the corresponding game tile.

See TileId::base() for more information on the difference between base and full tile identifiers.

Examples

let info = TileInfo::get(tile!(103, a, 2)).unwrap();
assert_eq!(info.base_id(), TileId::new(103, 1, 0));

pub fn full_id(&self) -> TileId

Full identifier of the corresponding game tile. If the tile has multiple graphical variants, this will return the first possible variant.

See TileId::base() for more information on the difference between base and full tile identifiers.

Examples

let info = TileInfo::get(tile!(103, b)).unwrap();
assert_eq!(info.full_id(), TileId::new(103, 2, 1));

pub fn iter() -> Iter<'static, Self>

Returns an interator over all tiles in the internal list.

pub fn get(id: TileId) -> Option<&'static Self>

Lookup tile information by tile identifier. The identifier can be in base or full format (with or without graphical variant).

pub fn count(&self) -> usize

Number of graphical variants that are available for a tile.

Examples

let info = TileInfo::get(tile!(103, b)).unwrap();
assert_eq!(info.count(), 3);

pub fn terrain(&self) -> Terrain

Terrain information for a tile.

Examples

let info = TileInfo::get(tile!(103, a)).unwrap();
assert_eq!(info.terrain(), Terrain::Asphalt(1));

let info = TileInfo::get(tile!(205, b)).unwrap();
assert_eq!(info.terrain(), Terrain::Gravel(2));

pub fn connection(&self, dir: Direction) -> Connection

Connection information for one of the six directions of a tile.

Examples

let info = TileInfo::get(tile!(103, b)).unwrap();
assert_eq!(info.connection(Direction::D), Connection::Straight(0));

let info = TileInfo::get(tile!(105, a)).unwrap();
assert_eq!(info.connection(Direction::D), Connection::Right(1));

pub fn connection_target(&self, source: Direction) -> Option<Direction>

Connection target direction for one of the six directions of a tile.

Follows the inner connection of a tile from the given source direction to the target. If the connection contains a Connection::Junction the first of both target directions is returned. If the connection does not contain a connection, None is returned.

Examples

let info = TileInfo::get(tile!(103, b)).unwrap();
assert_eq!(info.connection_target(Direction::D), Some(Direction::A));

let info = TileInfo::get(tile!(105, a)).unwrap();
assert_eq!(info.connection_target(Direction::D), Some(Direction::B));

pub fn edge(&self, dir: Direction) -> Edge

Edge informantion for one of the six directions of a tile.

Examples

let info = TileInfo::get(tile!(103, b)).unwrap();
assert_eq!(info.edge(Direction::D), Edge::Straight(2));

let info = TileInfo::get(tile!(105, a)).unwrap();
assert_eq!(info.edge(Direction::D), Edge::Straight(3));

pub fn series(&self) -> Series

Series information for a tile.

Examples

let info = TileInfo::get(tile!(103, a)).unwrap();
assert_eq!(info.series(), Series::Gt);

let info = TileInfo::get(tile!(205, b)).unwrap();
assert_eq!(info.series(), Series::Dirt);