Module rekee::token

Token which can be added to tiles. Can be used to add effects like a jump or water hazard area to a track tile. They are also used to indicate entrance (start) or exit of a joker section in rally-cross (RX) circuits.

Structs

EditionSummary

Edition and token count information for a specific edition.

TokenSummary

Count information for a specific base token.

Enums

ParseTokenError

TokenId

Identifier for a token which can be placed on tiles.

Traits

TokenList

Utility methods on any list of TokenIds.