Crate rekee

Library providing track editor functions for the Rekee web application.

When the library is compiled into a WASM module it will include a main function which connects the view components to the web page environment.

The following HTML elements are looked up by element id:

catalog-container (mandatory)

catalog-config-container (mandatory)

export-container (mandatory)

map-container (mandatory)

track-info-container (optional)

map-config-container (optional)

map-detail-container (optional)

version (optional)

welcome (optional)

The library can also be used outside the web environment as a native library to process track files. For details take a look at the files in the examples folder.

Modules

edition

Set of tiles that belong to the same game edition (core box or expansion).

hexagon

Types and methods for working with a hexagon grid.

import

Import and export of map data.

map

Map of tiles within a hexagon grid.

tile

Tile identifier and characteristics.

token

Token which can be added to tiles. Can be used to add effects like a jump or water hazard area to a track tile. They are also used to indicate entrance (start) or exit of a joker section in rally-cross (RX) circuits.

Macros

debug

Prints a debug message to the log output.

info

Prints an info message to the log output.

log

Prints a generic message to the log output.

tile

Creates a new tile identifier.

warn

Prints a warning message to the log output.