Enum rekee::edition::Edition

pub enum Edition {
    GtCoreBox,
    GtChampionship,
    GtWorldTour,
    GtTeamChallenge,
    GtAdrenalinePack,
    DirtCoreBox,
    Dirt110Percent,
    DirtRx,
    DirtClimb,
    DirtCopilot,
}

Rallyman game edition (core box or one of the expansions).

Variants

GtCoreBox

Rallyman: GT core box

GtChampionship

Championship expansion for Rallyman: GT

GtWorldTour

World Tour expansion for Rallyman: GT

GtTeamChallenge

Team Challenge expansion for Rallyman: GT

GtAdrenalinePack

Adrenaline Pack expansion for Rallyman: GT

DirtCoreBox

Rallyman: DIRT core box

Dirt110Percent

110% expansion for Rallyman: DIRT

DirtRx

RX expansion for Rallyman: DIRT

DirtClimb

Climb expansion for Rallyman: DIRT

DirtCopilot

Copilot expansion for Rallyman: DIRT

Implementations

impl Edition

pub fn all_tiles() -> Vec<TileId>

Returns all the tiles of the game joined into a sorted list.

Examples

let tiles = Edition::all_tiles();
assert_eq!(tiles.first(), Some(&tile!(101)));
assert_eq!(tiles.last(), Some(&tile!(905, b)));
assert_eq!(tiles.len(), 307);

pub fn gt_tiles() -> Vec<TileId>

Returns tiles part of Rallyman: GT game editions joined into a sorted list.

Examples

let tiles = Edition::gt_tiles();
assert_eq!(tiles.first(), Some(&tile!(101)));
assert_eq!(tiles.last(), Some(&tile!(901, b)));
assert_eq!(tiles.len(), 145);

pub fn dirt_tiles() -> Vec<TileId>

Returns tiles part of Rallyman: DIRT game editions joined into a sorted list.

Examples

let tiles = Edition::dirt_tiles();
assert_eq!(tiles.first(), Some(&tile!(201, a)));
assert_eq!(tiles.last(), Some(&tile!(905, b)));
assert_eq!(tiles.len(), 162);

pub fn tiles(&self) -> Vec<TileId>

Returns the tiles of a specific game edition.

Examples

let tiles = Edition::GtCoreBox.tiles();
assert_eq!(tiles.first(), Some(&tile!(101)));
assert_eq!(tiles.last(), Some(&tile!(120, b)));
assert_eq!(tiles.len(), 63);

pub fn contains(&self, tile: TileId) -> bool

Check whether a specific game edition contains the given tile.

Returns true when the identifier is found in the list of tiles.

Examples

let edition = Edition::DirtCoreBox;
assert_eq!(edition.contains(tile!(201, a)), true);
assert_eq!(edition.contains(tile!(325, b)), false);

let edition = Edition::DirtClimb;
assert_eq!(edition.contains(tile!(201, a)), false);
assert_eq!(edition.contains(tile!(325, b)), true);

pub fn iter() -> Iter<'static, Self>

Iterator over all game editions.