Enum rekee::hexagon::Direction

pub enum Direction {
    A,
    B,
    C,
    D,
    E,
    F,
}

Direction within a hexagon grid.

Used to indicate one of the six possible moving directions within the grid and to specify the rotation angle of hexagon tiles.

Examples

let dir = Direction::A;
assert_eq!(dir, 0.into());
assert_eq!(dir + Direction::B, Direction::B);
assert_eq!(dir - Direction::B, Direction::F);

Variants

A

Direction of the positive q axis.

B

Direction of the positive r axis.

C

Direction of the positive s axis.

D

Direction of the negative q axis.

E

Direction of the negative r axis.

F

Direction of the negative s axis.

Implementations

impl Direction

pub fn iter() -> Iter<'static, Self>

Iterator over all six directions.

Examples

let mut dirs = Direction::iter();
assert_eq!(dirs.next(), Some(&Direction::A));
assert_eq!(dirs.next(), Some(&Direction::B));
let mut dirs = dirs.skip(3);
assert_eq!(dirs.next(), Some(&Direction::F));
assert_eq!(dirs.next(), None);

pub fn opposite(&self) -> Self

Get opposite direction (point reflection).

Examples

let dir = Direction::A;
assert_eq!(dir.opposite(), Direction::D);

let dir = Direction::F;
assert_eq!(dir.opposite(), Direction::C);

pub fn rotated_left(&self) -> Self

Get direction rotated left (counter-clockwise).

Examples

let dir = Direction::B;
assert_eq!(dir.rotated_left(), Direction::A);

let dir = Direction::A;
assert_eq!(dir.rotated_left(), Direction::F);

pub fn rotated_right(&self) -> Self

Get direction rotated right (clockwise).

Examples

let dir = Direction::A;
assert_eq!(dir.rotated_right(), Direction::B);

let dir = Direction::F;
assert_eq!(dir.rotated_right(), Direction::A);

pub fn to_angle(&self, layout: &Layout) -> f32

Convert this hexagon grid direction into an angle in degrees.

Applies start_angle of the given layout. The retuned value is wrapped within the range of [0 .. 360) degrees.

Examples

let layout = Layout::new(Orientation::pointy(), Point(1.0, 1.0), Point(0.0, 0.0));

let dir = Direction::A;
assert_eq!(dir.to_angle(&layout), 90.0);

let dir = Direction::B;
assert_eq!(dir.to_angle(&layout), 150.0);