Module rekee::tile

Tile identifier and characteristics.

Structs

DangerLevelSummary
Tile count information for a specific terrain danger level.

EditionSummary
Edition and tile count information for a specific edition.

GroupByDangerLevel
An iterator that yields the terrain danger levels that are included in a list of tiles.

GroupByEdition
An iterator that yields the editions that are necessary to assemble a list of tiles.

GroupBySurface
An iterator that yields the terrain surfaces that are included in a list of tiles.

SurfaceSummary
Tile count information for a specific terrain surface.

TileId
Identifier of a game tile.

TileInfo
Information about tile characteristics like graphical variants, connections, and edge types.

Enums

Connection
Internal tile connection information. Describes how a specific tile edge is connected to other edges of the same tile.

ConnectionHint
Tile connection hint. Used to describe the orientation of adjacent tiles along the track in Map::append().

Edge
External tile edge information. Describes how the track spaces of the edge are connected to adjacent tiles.

ParseHintError

ParseTerrainError

Terrain
Tile terrain information. Describes the track surface of a tile together with the danger level.

Traits

TileList
Utilitiy methods on any list of TileIds.