rekee::tile

Trait TileList

pub trait TileList {
    // Required methods
    fn group_by_edition(&self) -> GroupByEdition ;
    fn group_by_terrain(&self) -> GroupByTerrain ;
    fn group_by_danger_level(&self) -> GroupByDangerLevel ;
    fn tile_summary(&self) -> Vec<TileSummary>;
    fn detect_series(&self) -> Option<Series>;

    // Provided methods
    fn edition_summary(&self) -> Vec<EditionSummary> { ... }
    fn terrain_summary(&self) -> Vec<TerrainSummary> { ... }
    fn danger_level_summary(&self) -> Vec<DangerLevelSummary> { ... }
}

Utility methods on any list of TileIds.

Required Methods

fn group_by_edition(&self) -> GroupByEdition

Returns an iterator over the editions that are necessary to build the current list of tiles.

Editions can occur multiple times during the iteration if more instances of a tile are used than what is provided by a single edition. Only the tile catalog number num is relevant for matching tiles and editions, side and var are ignored.

The order of editions returned by the iterator depends on the internal implementation.

Upon each step of the iterator the group of tiles that belong to the current edition is available in tiles.

Examples

let tiles = vec![tile!(101), tile!(103, a, 1), tile!(124, a), tile!(124, b)];
let mut iter = tiles.group_by_edition();

assert_eq!(iter.next(), Some(Edition::GtCoreBox));
assert_eq!(iter.tiles(), &[tile!(101), tile!(103, a, 1)][..]);
assert_eq!(iter.next(), Some(Edition::GtChampionship));
assert_eq!(iter.tiles(), &[tile!(124, b)][..]);
assert_eq!(iter.next(), Some(Edition::GtWorldTour));
assert_eq!(iter.tiles(), &[tile!(124, a)][..]);
assert_eq!(iter.next(), None);
assert_eq!(iter.tiles(), &[][..]);

Remaining tiles that do not belong to any edition are listed after the iterator has completed:

let tiles = vec![tile!(101), tile!(999)];
let mut iter = tiles.group_by_edition();

assert_eq!(iter.next(), Some(Edition::GtCoreBox));
assert_eq!(iter.tiles(), &[tile!(101)][..]);
assert_eq!(iter.next(), None);
assert_eq!(iter.tiles(), &[tile!(999)][..]);

fn group_by_terrain(&self) -> GroupByTerrain

Returns an iterator over the terrain surfaces that are included in the current list of tiles.

The order of terrain surfaces returned by the iterator depends on the internal implementation.

Upon each step of the iterator the group of tiles that belong to the current terrain is available in tiles.

Examples

let tiles = vec![tile!(301, a), tile!(220, b), tile!(418, a), tile!(419, b)];
let mut iter = tiles.group_by_terrain();

assert_eq!(iter.next(), Some(Terrain::Asphalt));
assert_eq!(iter.tiles(), &[tile!(301, a), tile!(418, a)][..]);
assert_eq!(iter.next(), Some(Terrain::Gravel));
assert_eq!(iter.tiles(), &[tile!(220, b), tile!(419, b)][..]);
assert_eq!(iter.next(), None);

Tiles that do not have terrain information are listed at the end, separated from filler tiles:

let tiles = vec![tile!(101), tile!(999, a)];
let mut iter = tiles.group_by_terrain();

assert_eq!(iter.next(), Some(Terrain::None));
assert_eq!(iter.tiles(), &[tile!(101)][..]);
assert_eq!(iter.next(), None);
assert_eq!(iter.tiles(), &[tile!(999, a)][..]);

fn group_by_danger_level(&self) -> GroupByDangerLevel

Returns an iterator over the danger levels that occur in the current list of tiles.

The order of danger levels returned by the iterator depends on the internal implementation.

Upon each step of the iterator the group of tiles that belong to the current danger level is available in tiles.

Examples

let tiles = vec![tile!(301, a), tile!(221, b), tile!(418, a), tile!(119, b)];
let mut iter = tiles.group_by_danger_level();

assert_eq!(iter.next(), Some(DangerLevel::Low));
assert_eq!(iter.tiles(), &[tile!(301, a), tile!(221, b)][..]);
assert_eq!(iter.next(), Some(DangerLevel::Medium));
assert_eq!(iter.tiles(), &[tile!(119, b)][..]);
assert_eq!(iter.next(), Some(DangerLevel::High));
assert_eq!(iter.tiles(), &[tile!(418, a)][..]);
assert_eq!(iter.next(), None);

Tiles that do not have terrain information are listed at the end, separated from filler tiles:

let tiles = vec![tile!(101), tile!(999, a)];
let mut iter = tiles.group_by_danger_level();

assert_eq!(iter.next(), Some(DangerLevel::None));
assert_eq!(iter.tiles(), &[tile!(101)][..]);
assert_eq!(iter.next(), None);
assert_eq!(iter.tiles(), &[tile!(999, a)][..]);

fn tile_summary(&self) -> Vec<TileSummary>

Returns a summary about the tiles that are necessary to build the current list of tiles.

For each base tile the entry contains the number of tile instances that are included in the current list of tiles. The returned array is always sorted by tile number.

Examples

let tiles = vec![tile!(101), tile!(103, a, 1), tile!(103, a, 2), tile!(124, a)];
for row in tiles.tile_summary() {
    println!("{}x tile {}", row.count, row.tile);
}

fn detect_series(&self) -> Option<Series>

Detects whether the given list of tiles belongs to the Rallyman GT or DIRT series.

Start/Finish tiles are considered first when detecting the game series. When no such tiles are found the tiles are checked whether they all belong to the same series.

Returns Some(Series) when detection was successful, otherwise returns None.

Provided Methods

fn edition_summary(&self) -> Vec<EditionSummary>

Returns a summary about the editions that are necessary to build the current list of tiles.

For each edition the entry contains the number of edition instances (boxes) required, and the number of tiles that belong to the edition.

This method is a convenience wrapper around the group_by_edition method. It merges the tile count of editions that occur multiple times. Different from group_by_edition the returned array is always sorted by edition. If the tile list contains unknown tiles an entry with edition set to None is put at the end of the summary.

Examples

let tiles = vec![tile!(101), tile!(103, a, 1), tile!(124, a), tile!(124, b)];
for row in tiles.edition_summary() {
    let label = match row.edition {
        Some(val) => val.to_string(),
        None => "Unknown".to_string(),
    };
    println!("{}: {}x ({} tiles)", label, row.edition_count, row.tile_count);
}

fn terrain_summary(&self) -> Vec<TerrainSummary>

Returns a summary about the terrain surfaces that occur within the current list of tiles.

For each terrain surface the entry contains the number of matching tiles.

This method is a convenience wrapper around the group_by_terrain method. Different from group_by_terrain the returned array is always sorted by terrain. If the tile list contains unknown tiles an entry with terrain set to None is put at the end of the summary.

Examples

let tiles = vec![tile!(301, a), tile!(220, b), tile!(418, a), tile!(419, b)];
for row in tiles.terrain_summary() {
    let label = match row.terrain {
        Some(val) => val.to_string(),
        None => "Unknown".to_string(),
    };
    println!("{}: {} tiles", label, row.tile_count);
}

fn danger_level_summary(&self) -> Vec<DangerLevelSummary>

Returns a summary about the danger levels that occur within the current list of tiles.

For each danger level the entry contains the number of matching tiles.

This method is a convenience wrapper around the group_by_danger_level method. Different from group_by_danger_level the returned array is always sorted by level value. If the tile list contains unknown tiles an entry with danger_level set to None is put at the end of the summary.

Examples

let tiles = vec![tile!(301, a), tile!(220, b), tile!(418, a), tile!(419, b)];
for row in tiles.danger_level_summary() {
    let label = match row.danger_level {
        Some(val) => val.to_string(),
        None => "Unknown".to_string(),
    };
    println!("{}: {} tiles", label, row.tile_count);
}