Trait rekee::tile::TileList

pub trait TileList {
    fn group_by_edition(&self) -> GroupByEdition;

    fn group_by_surface(&self) -> GroupBySurface;
}

Utilitiy methods on any list of TileIds.

Required methods

fn group_by_edition(&self) -> GroupByEdition

Returns an iterator over the editions that are necessary to build the current list of tiles.

Editions can occur multiple times during the iteration if more instances of a tile are used than what is provided by a single edition. Only the tile catalog number num is relevant for matching tiles and editions, side and var are ignored.

The order of editions returned by the iterator depends on the internal implementation.

Upon each step of the iterator the group of tiles that belong to the current edition is available in tiles.

Examples

let tiles = vec![tile!(101), tile!(103, a, 1), tile!(124, a), tile!(124, b)];
let mut iter = tiles.group_by_edition();

assert_eq!(iter.next(), Some(Edition::GtCoreBox));
assert_eq!(iter.tiles(), &[tile!(101), tile!(103, a, 1)][..]);
assert_eq!(iter.next(), Some(Edition::GtChampionship));
assert_eq!(iter.tiles(), &[tile!(124, b)][..]);
assert_eq!(iter.next(), Some(Edition::GtWorldTour));
assert_eq!(iter.tiles(), &[tile!(124, a)][..]);
assert_eq!(iter.next(), None);
assert_eq!(iter.tiles(), &[][..]);

Remaining tiles that do not belong to any edition are listed after the iterator has completed:

let tiles = vec![tile!(101), tile!(999)];
let mut iter = tiles.group_by_edition();

assert_eq!(iter.next(), Some(Edition::GtCoreBox));
assert_eq!(iter.tiles(), &[tile!(101)][..]);
assert_eq!(iter.next(), None);
assert_eq!(iter.tiles(), &[tile!(999)][..]);

fn group_by_surface(&self) -> GroupBySurface

Returns an iterator over the terrain surfaces that are included in the current list of tiles.

Only the tile terrain surface value is relevant for grouping of tiles, the danger_level is ignored. The order of terrain surfaces returned by the iterator depends on the internal implementation.

Upon each step of the iterator the group of tiles that belong to the current surface is available in tiles.

Examples

let tiles = vec![tile!(301, a), tile!(220, b), tile!(418, a), tile!(419, b)];
let mut iter = tiles.group_by_surface();

assert_eq!(iter.next(), Some(Terrain::Asphalt(0)));
assert_eq!(iter.tiles(), &[tile!(301, a), tile!(418, a)][..]);
assert_eq!(iter.next(), Some(Terrain::Gravel(0)));
assert_eq!(iter.tiles(), &[tile!(220, b), tile!(419, b)][..]);
assert_eq!(iter.next(), None);

Tiles that do not have terrain information are grouped with filler tiles:

let tiles = vec![tile!(101), tile!(999)];
let mut iter = tiles.group_by_surface();

assert_eq!(iter.next(), Some(Terrain::None));
assert_eq!(iter.tiles(), &[tile!(101), tile!(999)][..]);
assert_eq!(iter.next(), None);

Loading content...