Enum rekee::token::TokenId

pub enum TokenId {
    Chicane(Terrain),
    ChicaneWithLimit(Terrain),
    Jump(Terrain),
    Water(Terrain),
    ClimbAscent,
    ClimbDescent,
    Cloud,
    Oxygen(u8),
    JokerEntrance,
    JokerExit,
    Finish,
}

Identifier for a token which can be placed on tiles.

Examples

let token: TokenId = "chicane".parse().unwrap();
assert_eq!(token, TokenId::Chicane(Terrain::None));

let token = TokenId::Jump(Terrain::Gravel);
assert_eq!(token.to_string(), "Jump Gravel");

Variants

Chicane(Terrain)

Chicane token without a speed limit, part of the Rallyman DIRT Copilot Pack.

ChicaneWithLimit(Terrain)

Chicane token with a speed limit, part of the Rallyman DIRT Copilot Pack.

Jump(Terrain)

Jump token, part of the Rallyman DIRT Copilot Pack.

Water(Terrain)

Water token, part of the Rallyman DIRT Copilot Pack.

ClimbAscent

Ascent tile modifier, part of the Climb expansion for Rallyman DIRT.

ClimbDescent

Descent tile modifier, part of the Climb expansion for Rallyman DIRT.

Cloud

Cloud tile modifier, part of the Climb expansion for Rallyman DIRT.

Oxygen(u8)

Oxygen tile modifier, part of the Climb expansion for Rallyman DIRT.

JokerEntrance

Joker entrance arch, part of the RX expansion for Rallyman DIRT.

JokerExit

Joker exit arch, part of the RX expansion for Rallyman DIRT.

Finish

Finishing line arch, part of the RX expansion for Rallyman DIRT.

Implementations

impl TokenId

pub fn with_terrain(&self, terrain: Terrain) -> Self

Creates a copy of the current token with inner terrain changed to a given value.

Examples

let token = TokenId::Chicane(Terrain::Asphalt);
let other_token = token.with_terrain(Terrain::Snow);
assert_eq!(other_token, TokenId::Chicane(Terrain::Snow));

pub fn without_terrain(&self) -> Self

👎 Deprecated:
Please use TokenId::base() instead

pub fn base(&self) -> Self

Creates a copy of the current token without terrain info or oxygen level.

pub fn iter() -> Iter<'static, Self>

Iterator over all track tokens.

Tokens with terrain use the neutral Terrain::None value.

Examples

let tokens: Vec<String> = TokenId::iter()
    .map(|val| val.to_string())
    .collect();
assert_eq!(tokens, vec![
    "Chicane",
    "Chicane Limit",
    "Jump",
    "Water",
    "Climb Ascent",
    "Climb Descent",
    "Cloud",
    "Oxygen",
    "Joker Entrance",
    "Joker Exit",
    "Finish"]);