rekee::token

Trait TokenList

pub trait TokenList {
    // Required methods
    fn token_summary(&self) -> Vec<TokenSummary>;
    fn edition_summary(&self) -> Vec<EditionSummary>;
}

Utility methods on any list of TokenIds.

Required Methods

fn token_summary(&self) -> Vec<TokenSummary>

Returns a summary about the tokens that are contained within the current list of tokens.

For each token the entry contains the number of instances that occur in the current list of tiles. Note that terrain information is ignored when building the token summary. The returned array is always sorted by token identifier.

Examples

let tokens = vec![
    TokenId::JokerEntrance,
    TokenId::Chicane(Terrain::Gravel),
    TokenId::Chicane(Terrain::Asphalt),
    TokenId::Oxygen(1),
];
for row in tokens.token_summary() {
    println!("{}x token {}", row.count, row.token);
}

fn edition_summary(&self) -> Vec<EditionSummary>

Returns a summary about the editions that are necessary to build the current list of tokens.

For each edition the entry contains the number of tokens that belong to the edition.

Examples

let tokens = vec![
    TokenId::JokerEntrance,
    TokenId::Chicane(Terrain::Gravel),
    TokenId::Chicane(Terrain::Asphalt),
    TokenId::Oxygen(1),
];
for row in tokens.edition_summary() {
    println!("{}: {} tokens", row.edition, row.token_count);
}